api funcional

+++ IDGS903_Api/Models/compras.cs
@@ -7,8 +7,8 @@ namespace IDGS903_Api.Models
 
         [Key]
         public int Id { get; set; }
-        public float Cantidad { get; set; }
-        public float Total { get; set; }
+        public Double Cantidad { get; set; }
+        public Double Total { get; set; }
         public DateTime Fecha { get; set; }
         public int MateriaPrima_id { get; set; }
         public int Proveedor_id { get; set; }
